+++ akka-persistence-typed/src/main/scala/akka/persistence/typed/javadsl/Effect.scala
@@ -6,6 +6,7 @@ package akka.persistence.typed.javadsl
 import akka.annotation.DoNotInherit
 import akka.persistence.typed.scaladsl.PersistentBehaviors._
 import akka.japi.{ function ⇒ japi }
+import akka.persistence.typed.internal._
 
 import scala.collection.JavaConverters._
