+++ akka-persistence-typed/src/main/scala/akka/persistence/typed/javadsl/EventHandler.scala
@@ -8,7 +8,7 @@ import java.util.function.BiFunction
 import akka.util.OptionVal
 
 /**
- * SAM for reacting on events having been persisted
+ * FunctionalInterface for reacting on events having been persisted
  *
  * Used with [[EventHandlerBuilder]] to setup the behavior of a [[PersistentBehavior]]
  */
