+++ akka-persistence-typed/src/main/scala/akka/persistence/typed/scaladsl/PersistentBehaviors.scala
@@ -6,8 +6,7 @@ package akka.persistence.typed.scaladsl
 import akka.actor.typed.Behavior.UntypedBehavior
 import akka.actor.typed.scaladsl.ActorContext
 import akka.annotation.{ DoNotInherit, InternalApi }
-import akka.persistence.typed.internal.PersistentActorImpl
-import akka.persistence.typed.{ javadsl ⇒ j }
+import akka.persistence.typed.internal._
 
 import scala.collection.{ immutable ⇒ im }
 
@@ -73,7 +72,7 @@ object PersistentBehaviors {
    * Not for user extension.
    */
   @DoNotInherit
-  sealed trait Effect[+Event, State] {
+  trait Effect[+Event, State] {
     self: EffectImpl[Event, State] ⇒
     /* All events that will be persisted in this effect */
     def events: im.Seq[Event]
@@ -93,66 +92,11 @@ object PersistentBehaviors {
       CompositeEffect(this, Effect.stop[Event, State])
   }
 
-  // FIXME move to internal package?
-  /**
-   * INTERNAL API
-   */
-  @InternalApi
-  private[akka] sealed abstract class EffectImpl[+Event, State] extends j.Effect[Event, State] with Effect[Event, State] {
-    /* All events that will be persisted in this effect */
-    override def events: im.Seq[Event] = Nil
-
-    /* All side effects that will be performed in this effect */
-    override def sideEffects[E >: Event]: im.Seq[ChainableEffect[E, State]] = Nil
-  }
-
-  /**
-   * INTERNAL API
-   */
-  @InternalApi
-  private[akka] object CompositeEffect {
-    def apply[Event, State](effect: EffectImpl[Event, State], sideEffects: ChainableEffect[Event, State]): EffectImpl[Event, State] = {
-      CompositeEffect[Event, State](
-        effect,
-        sideEffects :: Nil
-      )
-    }
-  }
-
-  @InternalApi
-  private[akka] final case class CompositeEffect[Event, State](
-    persistingEffect: EffectImpl[Event, State],
-    _sideEffects:     im.Seq[ChainableEffect[Event, State]]) extends EffectImpl[Event, State] {
-
-    override val events = persistingEffect.events
-
-    override def sideEffects[E >: Event]: im.Seq[ChainableEffect[E, State]] = _sideEffects.asInstanceOf[im.Seq[ChainableEffect[E, State]]]
-
-  }
-
-  @InternalApi
-  private[akka] case object PersistNothing extends EffectImpl[Nothing, Nothing]
-
-  @InternalApi
-  private[akka] case class Persist[Event, State](event: Event) extends EffectImpl[Event, State] {
-    override def events = event :: Nil
-  }
-  @InternalApi
-  private[akka] case class PersistAll[Event, State](override val events: im.Seq[Event]) extends EffectImpl[Event, State]
-
   /**
    * Not for user extension
    */
   @DoNotInherit
-  sealed abstract class ChainableEffect[Event, State] extends EffectImpl[Event, State]
-
-  @InternalApi
-  private[akka] case class SideEffect[Event, State](effect: State ⇒ Unit) extends ChainableEffect[Event, State]
-  @InternalApi
-  private[akka] case object Stop extends ChainableEffect[Nothing, Nothing]
-
-  @InternalApi
-  private[akka] case object Unhandled extends EffectImpl[Nothing, Nothing]
+  abstract class ChainableEffect[Event, State] extends EffectImpl[Event, State]
 
   type CommandHandler[Command, Event, State] = (ActorContext[Command], State, Command) ⇒ Effect[Event, State]
